spark_jobs/requirements.txt
pyspark==3.5.3
python-dotenv
minio
redis
numpy
pymongo